// src/types/advertise.ts
export type ApiAdminAdvertise = {
  id: string;
  name: string;
  image: string;
  link?: string;
  is_enabled: boolean;
};

export type ApiAdvertise = {
  id: string;
  name: string;
  image: string;
  link?: string;
};

export type AdminAdvertise = {
  id: string;
  name: string;
  image: string;
  link?: string;
  isEnabled: boolean;
};

export type Advertise = {
  id: string;
  name: string;
  image: string;
  link?: string;
};